Added Storage, Parser, TaskList and Ui classes to make it more OOP
